requirements.txt
Flask==2.0.1
Flask_Login==0.5.0
Flask_QRcode==3.0.0
pony==0.7.14
python-dotenv==0.19.2
web3==5.25.0
